Fixes related to refactoring
